Move ConnectingFragment into ui
Complete connecting procedure
